+++ plugins/GPUSensors/GeforceSensors/nouveau_therm.cpp
@@ -228,7 +228,7 @@ int nouveau_therm_fan_rpm_get(struct nouveau_device *device)
     } while (cycles < stop && interval < 500000000);
     
     if (interval) {
-      return ((u64)60000000000LL * (((u64)cycles - 1) / 4)) / interval;
+      return (u32)(((u64)60000000000LL * (((u64)cycles - 1) / 4)) / interval);
     } else
       return 0;
   }
